Extracting work area methods for authenticated
The methods that I'm extracting are ones that make sense within the
context of an authenticated session.
